fixes to make examples work

+++ Code/functions.py
@@ -3,6 +3,7 @@ from collections import deque
 from scipy.special import erf
 from astropy.convolution import convolve, Box1DKernel, Gaussian1DKernel, convolve_fft
 from scipy.ndimage import filters
+import pdb
 
 ##########################################################################################
 #                                                                                        #
@@ -152,11 +153,11 @@ def mean_smooth_ind(x, y, width):
 
 
 def bin_data(x, y, params, log = True):
-
+    
     if log:
         mi = min(np.log10(x))
         ma = max(np.log10(x))
-        no = np.ceil((ma-mi)/params['findex']['binning'])
+        no = np.int(np.ceil((ma-mi)/params['findex']['binning']))
         bins = np.logspace(mi, mi+no*params['findex']['binning'], no)
     else:
         bins = np.arange(min(x), max(x)+params['findex']['binning'], params['findex']['binning'])
